+++ src/base/widget.c
@@ -64,7 +64,6 @@ static ret_t widget_on_paint_end(widget_t* widget, canvas_t* c);
 
 typedef widget_t* (*widget_find_wanted_focus_widget_t)(widget_t* widget, darray_t* all_focusable);
 static ret_t widget_move_focus(widget_t* widget, widget_find_wanted_focus_widget_t find);
-static widget_t* widget_get_real_window_or_keyboard(widget_t* widget);
 
 #define widget_set_xywh(widget, val, update_layout, invalidate)    \
   do {                                                             \
@@ -3673,7 +3672,7 @@ static widget_t* widget_get_window_or_keyboard(widget_t* widget) {
   return NULL;
 }
 
-static widget_t* widget_get_real_window_or_keyboard(widget_t* widget) {
+widget_t* widget_get_real_window_or_keyboard(widget_t* widget) {
   return_value_if_fail(widget != NULL && widget->vt != NULL, NULL);
 
   widget_t* win = widget_get_window_or_keyboard(widget);